feat(examples): introduce tests for example storage
