+++ examples/storage.rs
@@ -37,3 +37,36 @@ pub fn get() -> i32 {
 
 #[cfg(not(target_arch = "wasm32"))]
 fn main() {}
+
+#[test]
+fn test() -> anyhow::Result<()> {
+    use zint::{Bytes32, Contract, InstructionResult, U256};
+    let mut contract = Contract::search("storage")?.compile()?;
+
+    {
+        let key = 0;
+        let value: i32 = 42;
+        let info = contract.execute(&[b"set(i32)".to_vec(), value.to_bytes32().to_vec()])?;
+        assert!(info.ret.is_empty());
+        assert_eq!(info.instr, InstructionResult::Return);
+        assert_eq!(info.storage.get(&U256::from(key)), Some(&U256::from(value)));
+    }
+
+    {
+        let info = contract.execute(&["get()"])?;
+        assert_eq!(info.instr, InstructionResult::Return);
+        assert_eq!(info.ret, 0.to_bytes32());
+    }
+
+    {
+        let key = 0;
+        let value = 42;
+        let info =
+            contract.execute(&[b"set_and_get(i32)".to_vec(), value.to_bytes32().to_vec()])?;
+        assert_eq!(info.instr, InstructionResult::Return);
+        assert_eq!(info.ret, value.to_bytes32());
+        assert_eq!(info.storage.get(&U256::from(key)), Some(&U256::from(value)));
+    }
+
+    Ok(())
+}
